backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-binance==1.0.19
pandas==2.3.2
pandas-ta
apscheduler==3.10.4
firebase-admin==6.5.0
pydantic==2.9.0
httpx==0.27.0
pytest==8.3.3
pytest-asyncio==0.24.0
